% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATcor.R
\name{MATcor}
\alias{MATcor}
\title{MATcor: Matriz de correlación gráfica}
\usage{
MATcor(data, ...)
}
\arguments{
\item{data}{Dataframe que contiene los datos.}

\item{...}{Variables específicas (sin entrecomillar) para incluir en la matriz de correlación. Si no se especifica, se usan todas las variables métricas.}
}
\value{
Una lista con el gráfico de correlación, guardada en el entorno global con el nombre "<nombre_df>_correlaciones_info".
}
\description{
Esta función crea una matriz de correlación gráfica entre las variables métricas de un dataframe.
Por defecto, incluye todas las variables métricas del dataframe, pero también permite seleccionar
un subconjunto específico de variables.
}
\examples{
\u0026dontrun{
# Correlaciones entre todas las variables métricas del dataframe mtcars
MATcor(data = mtcars)

# Correlaciones entre un subconjunto específico de variables
MATcor(data = mtcars, mpg, hp, wt)
}
}
