% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATfexcel.R
\name{MATfexcel}
\alias{MATfexcel}
\title{MATfexcel: Importar datos desde una hoja de Excel}
\usage{
MATfexcel(file_path, sheet_name, na_values = NULL)
}
\arguments{
\item{file_path}{Ruta al archivo Excel.}

\item{sheet_name}{Nombre o índice de la hoja de Excel a importar.}

\item{na_values}{Un vector de caracteres que representa valores que deben tratarse como \code{NA}. Por defecto, las celdas vacías se consideran \code{NA}.}
}
\value{
Un dataframe con los datos importados.
}
\description{
Esta función importa datos desde una hoja de cálculo Excel, convierte la primera columna
en nombres de fila, y muestra un resumen de las variables.
}
\examples{
\dontrun{
# Importar datos con celdas vacías como NA
my_data <- MATfexcel("datos.xlsx", "Datos")

# Importar datos con valores adicionales considerados como NA
my_data <- MATfexcel("datos.xlsx", "Datos", na_values = c("n.d.", "s.d."))
}
}
