% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATout_Mahalanobis.R
\name{MATout_Mahalanobis}
\alias{MATout_Mahalanobis}
\title{Identificación de Outliers Multivariados mediante la Distancia de Mahalanobis}
\usage{
MATout_Mahalanobis(data, ...)
}
\arguments{
\item{data}{Dataframe que contiene los datos a analizar.}

\item{...}{Variables a incluir en el análisis (sin comillas, separadas por comas). Opcional.
Si no se indican variables, por defecto se utilizan todas las columnas numéricas del dataframe.}
}
\value{
La función crea dos objetos en el entorno global:
\itemize{
\item Un nuevo dataframe sin los outliers, nombrado como \code{<nombre_original>_so}.
\item Una lista de información nombrada como \code{<nombre_original>_so_info}, que incluye:
\itemize{
\item \code{Outliers_Table}: Una tabla en formato HTML con los outliers detectados.
\item \code{Boxplot}: Un gráfico de caja de las distancias de Mahalanobis.
}
}
}
\description{
Esta función identifica outliers multivariados utilizando la distancia de Mahalanobis.
Genera un dataframe sin los outliers y proporciona un gráfico de caja y una tabla con los casos detectados.
}
\examples{
\dontrun{
datos <- data.frame(x = rnorm(100), y = rnorm(100))
MATout_Mahalanobis(datos, x, y)
# Resultados: datos_so y datos_so_info en el entorno global

# Si no se indican variables, se usan todas las columnas numéricas:
MATout_Mahalanobis(datos)
# Resultados: datos_so y datos_so_info en el entorno global
}

}
