% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATmv.R
\name{MATmv}
\alias{MATmv}
\title{MATmv: Manejo de datos con casos completos y NA}
\usage{
MATmv(dataframe, columnas)
}
\arguments{
\item{dataframe}{Un dataframe que contiene las variables.}

\item{columnas}{Las columnas que se desean seleccionar.}
}
\value{
El nombre del nuevo dataframe con los casos completos.
}
\description{
Esta función selecciona variables de un dataframe, identifica casos completos y crea un resumen
con información sobre los datos faltantes.
}
